dcNotifyTg/etc/dcNotifyTg/requirements.txt
python-telegram-bot
sqlalchemy
requests